refact(wasm): use helper module instead of banner (#607)
* refact(wasm): use helper module instead of banner

* test(wasm): test inject helper logic
