+++ packages/wasm/src/index.ts
@@ -6,6 +6,8 @@ import { Plugin } from 'rollup';
 
 import { RollupWasmOptions } from '../types';
 
+import { getHelpersModule, HELPERS_ID } from './helper';
+
 export function wasm(options: RollupWasmOptions = {}): Plugin {
   const { sync = [], maxFileSize = 14 * 1024, publicPath = '' } = options;
 
@@ -15,7 +17,19 @@ export function wasm(options: RollupWasmOptions = {}): Plugin {
   return {
     name: 'wasm',
 
+    resolveId(id) {
+      if (id === HELPERS_ID) {
+        return id;
+      }
+
+      return null;
+    },
+
     load(id) {
+      if (id === HELPERS_ID) {
+        return getHelpersModule();
+      }
+
       if (!/\.wasm$/.test(id)) {
         return null;
       }
@@ -46,59 +60,6 @@ export function wasm(options: RollupWasmOptions = {}): Plugin {
       );
     },
 
-    banner: `
-      function _loadWasmModule (sync, filepath, src, imports) {
-        function _instantiateOrCompile(source, imports, stream) {
-          var instantiateFunc = stream ? WebAssembly.instantiateStreaming : WebAssembly.instantiate;
-          var compileFunc = stream ? WebAssembly.compileStreaming : WebAssembly.compile;
-          
-          if (imports) {
-            return instantiateFunc(source, imports)
-          } else {
-            return compileFunc(source)
-          }
-        }
-
-        var buf = null
-        var isNode = typeof process !== 'undefined' && process.versions != null && process.versions.node != null
-        
-        if (filepath && isNode) {
-          var fs = eval('require("fs")')
-          var path = eval('require("path")')
-
-          return new Promise((resolve, reject) => {
-            fs.readFile(path.resolve(__dirname, filepath), (error, buffer) => {
-              if (error != null) {
-                reject(error)
-              }
-
-              resolve(_instantiateOrCompile(buffer, imports, false))
-            });
-          });
-        } else if (filepath) {
-          return _instantiateOrCompile(fetch(filepath), imports, true)
-        }
-        
-        if (isNode) {
-          buf = Buffer.from(src, 'base64')
-        } else {
-          var raw = globalThis.atob(src)
-          var rawLength = raw.length
-          buf = new Uint8Array(new ArrayBuffer(rawLength))
-          for(var i = 0; i < rawLength; i++) {
-             buf[i] = raw.charCodeAt(i)
-          }
-        }
-
-        if(sync) {
-          var mod = new WebAssembly.Module(buf)
-          return imports ? new WebAssembly.Instance(mod, imports) : mod
-        } else {
-          return _instantiateOrCompile(buf, imports, false)
-        }
-      }
-    `.trim(),
-
     transform(code, id) {
       if (code && /\.wasm$/.test(id)) {
         const isSync = syncFiles.indexOf(id) !== -1;
@@ -115,7 +76,8 @@ export function wasm(options: RollupWasmOptions = {}): Plugin {
           src = null;
         }
 
-        return `export default function(imports){return _loadWasmModule(${+isSync}, ${publicFilepath}, ${src}, imports)}`;
+        return `import { _loadWasmModule } from ${JSON.stringify(HELPERS_ID)};
+export default function(imports){return _loadWasmModule(${+isSync}, ${publicFilepath}, ${src}, imports)}`;
       }
       return null;
     },
