+++ packages/wasm/test/fixtures/foo.js
@@ -1,0 +1,1 @@
+export default 'foo';
